Make it easier to replace SCMonitor implementation
