+++ src/main/java/com/saucelabs/ci/sauceconnect/SCMonitor.java
@@ -1,103 +1,9 @@
 package com.saucelabs.ci.sauceconnect;
 
-import org.json.JSONObject;
-import org.slf4j.Logger;
-
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
 import java.util.concurrent.Semaphore;
 
-/** Monitors SC Process via HTTP API */
-public class SCMonitor extends Thread {
-    private Semaphore semaphore;
-    private final int port;
-    private final Logger LOGGER;
-
-    private boolean failed;
-    private boolean apiResponse;
-
-    private HttpClient client = HttpClient.newHttpClient();
-    private static final int sleepTime = 1000;
-
-    public SCMonitor(
-        String name,
-        final int port,
-        final Logger logger) {
-        super(name);
-        this.port = port;
-        this.LOGGER = logger;
-    }
-
-    public void setSemaphore(Semaphore semaphore) {
-        this.semaphore = semaphore;
-    }
-
-    public String getTunnelId() {
-        HttpRequest request = HttpRequest.newBuilder()
-            .uri(URI.create(String.format("http://localhost:%d/info", port)))
-            .GET()
-            .build();
-        try {
-            HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
-            String responseBody = response.body();
-            JSONObject jsonObject = new JSONObject(responseBody);
-            if (jsonObject.has("tunnel_id")) {
-                return jsonObject.getString("tunnel_id");
-            }
-        } catch (Exception e) {
-            this.LOGGER.info("Failed to get tunnel id", e);
-            return null;
-        }
-        this.LOGGER.info("Failed to get tunnel id");
-        return null;
-    }
-
-    public boolean isFailed() {
-        return failed;
-    }
-
-    public void run() {
-        while (true) {
-            pollEndpoint();
-            if (this.semaphore.availablePermits() > 0) {
-                return;
-            }
-
-            try {
-                Thread.sleep(sleepTime);
-            } catch ( java.lang.InterruptedException e ) {
-                return;
-            }
-        }
-    }
-
-    protected void pollEndpoint() {
-        HttpRequest request = HttpRequest.newBuilder()
-            .uri(URI.create(String.format("http://localhost:%d/readyz", port)))
-            .GET()
-            .build();
-
-        try {
-            HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
-
-            if (response.statusCode() == 200) {
-                this.apiResponse = true;
-                this.LOGGER.info("Got connected status");
-                semaphore.release();
-            } else if (response.statusCode() == 503) {
-                this.apiResponse = true;
-            }
-        } catch ( Exception e ) {
-            if ( this.apiResponse ) {
-                // We've had a successful API endpoint read, but then it stopped responding, which means the process failed to start
-                this.failed = true;
-                this.LOGGER.warn("API stopped responding", e);
-                semaphore.release();
-            }
-        }
-
-        this.LOGGER.trace("No API response yet");
-    }
+public interface SCMonitor extends Runnable {
+    void setSemaphore(Semaphore semaphore);
+    String getTunnelId();
+    boolean isFailed();
 }
